requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
comtypes==1.1.11
iniconfig==1.1.1
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.7
pypiwin32==223
pytest==7.1.1
pyttsx3==2.90
pywin32==303
tomli==2.0.1
typing_extensions==4.1.1
